Properties: Extract TracesTab, add AxesTab
