+++ applications/databrowser/src/main/resources/org/csstudio/trends/databrowser3/messages.properties
@@ -147,14 +147,12 @@ LabelFontLbl=Label Font:
 LabelFontTT=Select font for axes labels
 LegendFontLbl=Legend Font:
 LegendFontTT=Select font for the legend
-LinacScale=Linear
-LinLogScaleType=Scale Type
 LiveBufferSizeInfoFmt=Assuming one value per second, {0} elements cover {1}
 LiveData=Live Data
 LiveSampleBufferSize=Buffer Size
 LogentryDefaultBody=See attached Data Browser plot
 LogentryDefaultTitle=Data Browser Plot
-LogScale=Logarithmic
+LogScale=Log. Scale
 Miscellaneous=Misc.
 Model_Disconnected=Disconnected
 MoveItemDown=Move Down
